content: move the main repo to GitHub
As much as I like sr.ht in some ways, very few people contribute on it
and maintaining all the extra stuff is just too much work.
Until we have enough contributors to decide more permanently, I'm moving
back to GitHub as the main repo.
Hopefully we can find (or start) a better, more cooperative, code forge
in the future.

Signed-off-by: Sam Whited <sam@samwhited.com>